requirements.txt
requests
python-dotenv
json
OpenAI
